Update View Components (#5)
* move Display in "View" Directory

* shell script for faster testing

* adjust test file

* update README
