+++ src/Model/Agents/Sheep.java
@@ -1,4 +1,5 @@
 package Model.Agents;
+
 import java.awt.Color;
 
 public class Sheep extends Agent{
